refactor page types, add EOT page types
