--- types/LandingPage.ts
@@ -1,17 +1,0 @@
-import { ImageTypes } from './Base';
-
-export type LandingPage = {
-    metadataTitle: string;
-    metadataDescription: string;
-    metadataSocialMediaImage: ImageTypes;
-    heroImage: ImageTypes;
-    heroTagline: any;
-    missionTitle: string;
-    missionDescription: any;
-    itemsCollection: any;
-    workForYouDescription: any;
-    workForYouImage: ImageTypes;
-    lifeAsATorchboxer: any;
-    ctaTitle: string;
-    ctaDescription: any;
-};
